backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
PyJWT
jinja2
python-dotenv
aiofiles
openpyxl
reportlab
